requirements.txt
fastapi
uvicorn
python-multipart
Pillow
numpy
torch
transformers
requests
Jinja2
aiofiles
supabase
tqdm
python-dotenv
psycopg2
mwclient